deletion authentication, removed from post page better back to home page button

+++ pages/api/delete_post/[delete_post].js
@@ -1,10 +1,20 @@
 import fs from 'fs'
+import { getSession } from 'next-auth/react'
+import permissions from '../../../user_permissions.json'
 
 export default async function handler(req, res) {
+    const session = await getSession({ req })
+    if(!session || !permissions.admins.includes(session.user.email)) {
+        return res.status(403).end()
+    }
     if(req.method !== 'DELETE') {
         return res.status(405).send('Only DELETE requests allowed!')
     }
     const { delete_post } = await req.query
-    fs.unlinkSync('./posts/' + delete_post + '.md')
-    res.status(204).end()
+    try {
+        fs.unlinkSync('./posts/' + delete_post + '.md')
+    } catch(e) {
+        return res.status(410).end()
+    }
+    return res.status(204).end()
 }
